изменила тип колонки users_type

+++ src/main/java/pro/sky/animal_shelter_ji22_team1_app/repository/UserRepository.java
@@ -9,5 +9,5 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
     List<UserEntity> findByLogin(String login);
     List<UserEntity> findByPhone(String phone);
 
-    UserEntity findByChatId(Long chatId);
+    UserEntity findByChatId(Integer chatId);
 }
